requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1

langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.5

# Интеграция с Gemini через LangChain
langchain-google-genai

chromadb==0.5.18
pypdf==5.1.0

uvloop==0.21.0